+++ lib/core/models/sum_of_task_time_model.dart
@@ -1,5 +1,5 @@
 class TaskByTaskModel {
   String taskName;
-  int passingTime;
+  double passingTime;
   TaskByTaskModel(this.taskName, this.passingTime);
 }
